Support for multiple radar sites
Refactor GisPluginRadar to use a separate object for radar sites.
Many of the plugin structure's fields are now included in a separate
RadarSite structure. Multiple such structures can (and are) added to the
viewer and automatically load/unload themselves when the view changes.

Also move config generation code to the Level2 object as well as
some slight improvements to it.