Add dual-pole support and improve colormap handling
Add support for dual-pole (KOUN). This should work for the most part.
Some of the dual-pole colormaps are still broken. For KOUN to work, a
recent (unreleased) version of the RSL wsr88d_locations.dat file is
needed.

Use non builtin colormaps. This should make updating colormaps easier.
The colormaps are written in simple flat files, the format is:

  - line 1: Parameter name,  - Reflectivity
  - line 2: Data scale,      - 2
  - line 3: Data shift,      - 31
  - line 4: RGBA color data, - 255 207 0 255
  - line n: Same as line 4 for however many lines are needed

The scale and shift parameters map the parameter values to a colortable
index. The index is clamped to the number of color data lines.