Switch back to only showing one radar at a time
Radars are still loaded automatically but only show the radar in the
foreground tab of the configuration pane. With multiple radars (and
checkboxes) it becomes a hassle to switch between them because it
requires checking one box, unchecking another, and switching the tab.

Also, showing multiple radars didn't really work because of how the
depth test was done.