Add new application icons
The new icon is works better with different background colors, it also
avoids using flat background colors or borders.

The small icons (less than 32) pixels do not includes the sun in the
background because it is difficult to see at such small sizes and just
add clutter.

The old icon is kept as logo.svg and is used in the about dialog. It may
be converted to use the new icon eventually as well.