Add warning/watch/alert plugin
Alert information is shown similar to what is shown on the main
weather.gov page. Alerts are retrieved in real time using the NWS CAP
feed [1] which includes most, but not all, of the information from
weather.gov. More details about the CAP feed can be found on the NWS CAP
Wiki [2].

Both county based warnings and polygon based warnings are support. In
addition, when zoomed in close enough county outlines will be drawn for
each state.

Additional information can be retrieved about a particular alert by
clicking on the highlighted area in the viewer.

Alerts can be hidden by toggling the alert type in the alert plugin's
configuration tab.

If you are developing from Git, you will need to download a FIPS county
polygons file [2]. This should be done automatically on the first build.
If it is not successful, the polygons are also included in the >= 0.6
tarballs.

There are currently several drawbacks to this plugin:

  - NWS warning zones are not use, only FIPS counties are used.

  - The county polygons have been simplified for performance reasons.

  - The NWS CAP feed does not include all the text products showed on
    the main weather.gov page. For example, hazardous weather outlooks
    are not included. This may be improved slightly by the NWS in future
    revision of the CAP feed.

  - Historic data is not available but previously downloaded alerts are
    cached and can be viewed.

[1] http://alerts.weather.gov/cap/us.php?x=0
[2] http://wiki.citizen.apps.gov/nws_developers/
[3] http://lug.rose-hulman.edu/proj/aweather/files/fips/