Update objects for new Grits expose settings
All objects are drawn with sorting (depth testing) disabled except for
the Level-II sweeps and volume scans. The order things appear is
determined based on the level they are drawn in since depth testing
works across levels now.

The order is as follows, from lowest to highest:

  WORLD+1 - Counties outlines     - no depth test
  WORLD+1 - County-based warnings - no depth test
  WORLD+2 - Conus radar overlays  - no depth test
  WORLD+3 - Level-II radar sweeps - depth test enabled
  WORLD+4 - Storm-based warnings  - no depth test
  WORLD+5 - Level-II Volume scans - depth test enabled
  OVERLAY - Level-II site markers - no depth test
  HUD     - Level-II color table  - no depth test

The advantage to this order is that:
  - County-based warnings do not effect the radar colors
  - Level-II sweeps do not hide storm-based warnings
  - Storm based warnings don't show through volume scans