Revert "Fix off-by-one on date"
This issue is related to whether or not the users timezone accounts for
leap seconds in Unix timestamps.

For instance, 1325376000 in TZ=UTC is 2011-01-01 00:00:00, but in
TZ=right/UTC it is 2011-12-31 23:59:36.

RSL generates the dates by using gmtime at time 00:00:00 on a particular
date since the epoch. If the timezone includes leap seconds, this will
return the previous day rather than the desired day, which is why the
original +1 was added to AWeather.

However, nobody uses timezones that include leap seconds (i.e. the
right/ folder) so it's better to leave it how it is.